+++ Sources/SwanKit/Storage/Resources.md
@@ -1,6 +1,15 @@
-## Optimization
+## CPU Optimization
+
+An entry level article:
 
 - https://www.appcoda.com/accelerate-framework/
+
+Framework which uses Accelerate library:
+
+- https://github.com/mattt/Surge
+
+C++ articles related to optimization and SIMD:
+
 - http://www.agner.org/optimize/
 - http://www.agner.org/optimize/optimizing_cpp.pdf
 - https://www.kernel.org/pub/linux/kernel/people/geoff/cell/ps3-linux-docs/CellProgrammingPrimer.html
